fix(logout): navbar should be updated when clicking 'Logout'
Before this fix, when clicking the 'Logout' button, the user would be
redirected to the login form, but the 'Logout' button would persist.
This happened because the navbar items were being determined when the
componented was being mounted (i.e., using the `mounted()` lifecycle
method of the component). Now, with this fix, the navbar items are
updated using the `beforeMount()` lifecycle method.